feat: show inputs table
