+++ app/cli/cmd/available_integration_describe.go
@@ -22,6 +22,7 @@ import (
 	"sort"
 
 	"github.com/chainloop-dev/chainloop/app/cli/internal/action"
+	"github.com/chainloop-dev/chainloop/app/controlplane/extensions/sdk/v1"
 	"github.com/jedib0t/go-pretty/v6/table"
 	"github.com/jedib0t/go-pretty/v6/text"
 	"github.com/spf13/cobra"
@@ -92,7 +93,7 @@ func availableIntegrationDescribeTableOutput(items []*action.AvailableIntegratio
 }
 
 // render de-normalized schema format
-func renderSchemaTable(tableTitle string, properties action.SchemaPropertiesMap) error {
+func renderSchemaTable(tableTitle string, properties sdk.SchemaPropertiesMap) error {
 	if len(properties) == 0 {
 		return nil
 	}
