+++ app/cli/internal/action/available_integration_list.go
@@ -16,13 +16,12 @@
 package action
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
-	"sort"
 
 	pb "github.com/chainloop-dev/chainloop/app/controlplane/api/controlplane/v1"
+	"github.com/chainloop-dev/chainloop/app/controlplane/extensions/sdk/v1"
 	"github.com/santhosh-tekuri/jsonschema/v5"
 )
 
@@ -45,22 +44,8 @@ type JSONSchema struct {
 	Raw string `json:"schema"`
 	// Parsed schema so it can be used for validation or other purposes
 	// It's not shown in the json output
-	Parsed     *jsonschema.Schema  `json:"-"`
-	Properties SchemaPropertiesMap `json:"-"`
-}
-
-type SchemaPropertiesMap map[string]*SchemaProperty
-type SchemaProperty struct {
-	// Name of the property
-	Name string
-	// optional description
-	Description string
-	// Type of the property (string, boolean, number)
-	Type string
-	// If the property is required
-	Required bool
-	// Optional format (email, host)
-	Format string
+	Parsed     *jsonschema.Schema      `json:"-"`
+	Properties sdk.SchemaPropertiesMap `json:"-"`
 }
 
 func NewAvailableIntegrationList(cfg *ActionsOpts) *AvailableIntegrationList {
@@ -107,24 +92,24 @@ func pbAvailableIntegrationItemToAction(in *pb.IntegrationAvailableItem) (*Avail
 
 	// Parse the schemas so they can be used for validation or other purposes
 	var err error
-	i.Registration.Parsed, err = compileJSONSchema(foType.GetRegistrationSchema())
+	i.Registration.Parsed, err = sdk.CompileJSONSchema(foType.GetRegistrationSchema())
 	if err != nil {
 		return nil, fmt.Errorf("failed to compile registration schema: %w", err)
 	}
 
-	i.Attachment.Parsed, err = compileJSONSchema(foType.GetAttachmentSchema())
+	i.Attachment.Parsed, err = sdk.CompileJSONSchema(foType.GetAttachmentSchema())
 	if err != nil {
 		return nil, fmt.Errorf("failed to compile registration schema: %w", err)
 	}
 
 	// Calculate the properties map
-	i.Registration.Properties = make(SchemaPropertiesMap)
-	if err := calculatePropertiesMap(i.Registration.Parsed, &i.Registration.Properties); err != nil {
+	i.Registration.Properties = make(sdk.SchemaPropertiesMap)
+	if err := sdk.CalculatePropertiesMap(i.Registration.Parsed, &i.Registration.Properties); err != nil {
 		return nil, fmt.Errorf("failed to calculate registration properties: %w", err)
 	}
 
-	i.Attachment.Properties = make(SchemaPropertiesMap)
-	if err := calculatePropertiesMap(i.Attachment.Parsed, &i.Attachment.Properties); err != nil {
+	i.Attachment.Properties = make(sdk.SchemaPropertiesMap)
+	if err := sdk.CalculatePropertiesMap(i.Attachment.Parsed, &i.Attachment.Properties); err != nil {
 		return nil, fmt.Errorf("failed to calculate attachment properties: %w", err)
 	}
 
@@ -132,84 +117,4 @@ func pbAvailableIntegrationItemToAction(in *pb.IntegrationAvailableItem) (*Avail
 	i.SubscribedInputs = append(i.SubscribedInputs, foType.GetSubscribedMaterials()...)
 
 	return i, nil
-}
-
-func compileJSONSchema(in []byte) (*jsonschema.Schema, error) {
-	// Parse the schemas
-	compiler := jsonschema.NewCompiler()
-	// Enable format validation
-	compiler.AssertFormat = true
-	// Show description
-	compiler.ExtractAnnotations = true
-
-	if err := compiler.AddResource("schema.json", bytes.NewReader(in)); err != nil {
-		return nil, fmt.Errorf("failed to compile schema: %w", err)
-	}
-
-	return compiler.Compile("schema.json")
-}
-
-// calculate a map with all the properties of a schema
-func calculatePropertiesMap(s *jsonschema.Schema, m *SchemaPropertiesMap) error {
-	if m == nil {
-		return nil
-	}
-
-	// Schema with reference
-	if s.Ref != nil {
-		return calculatePropertiesMap(s.Ref, m)
-	}
-
-	// Appended schemas
-	if s.AllOf != nil {
-		for _, s := range s.AllOf {
-			if err := calculatePropertiesMap(s, m); err != nil {
-				return err
-			}
-		}
-	}
-
-	if s.Properties != nil {
-		requiredMap := make(map[string]bool)
-		for _, r := range s.Required {
-			requiredMap[r] = true
-		}
-
-		for k, v := range s.Properties {
-			if err := calculatePropertiesMap(v, m); err != nil {
-				return err
-			}
-
-			var required = requiredMap[k]
-			(*m)[k] = &SchemaProperty{
-				Name:        k,
-				Type:        v.Types[0],
-				Required:    required,
-				Description: v.Description,
-				Format:      v.Format,
-			}
-		}
-	}
-
-	// We return the map sorted
-	// This is not strictly necessary but it makes the output more readable
-	// and it's easier to test
-
-	// Sort the keys
-	keys := make([]string, 0, len(*m))
-	for k := range *m {
-		keys = append(keys, k)
-	}
-
-	sort.Strings(keys)
-
-	// Create a new map with the sorted keys
-	newMap := make(SchemaPropertiesMap)
-	for _, k := range keys {
-		newMap[k] = (*m)[k]
-	}
-
-	*m = newMap
-
-	return nil
 }
