+++ app/controlplane/extensions/core/smtp/v1/README.md
@@ -27,6 +27,15 @@ Starting now, every time a workflow run occurs, an email notification will be se
 
 ## Registration Input Schema
 
+|Field|Type|Required|Description|
+|---|---|---|---|
+|from|string (email)|yes|The email address of the sender.|
+|host|string|yes|The host to use for the SMTP authentication.|
+|password|string|yes|The password to use for the SMTP authentication.|
+|port|string|yes|The port to use for the SMTP authentication|
+|to|string (email)|yes|The email address to send the email to.|
+|user|string|yes|The username to use for the SMTP authentication.|
+
 ```json
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
@@ -75,6 +84,10 @@ Starting now, every time a workflow run occurs, an email notification will be se
 
 ## Attachment Input Schema
 
+|Field|Type|Required|Description|
+|---|---|---|---|
+|cc|string (email)|no|The email address of the carbon copy recipient.|
+
 ```json
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
@@ -89,4 +102,4 @@ Starting now, every time a workflow run occurs, an email notification will be se
   "additionalProperties": false,
   "type": "object"
 }
-````
+```
