+++ app/controlplane/extensions/sdk/v1/testdata/schemas/basic.json
