+++ app/controlplane/extensions/sdk/v1/testdata/schemas/oneof_required.json
